beta need sprites and wall bug with mlx new image

+++ cube3D/wall.c
@@ -5,13 +5,13 @@
 **the correct distance to a wall
 */
 
-float walls(t_info *info, float distance, int whichray)
+double walls(t_info *info, double distance, int whichray)
 {
-	float corrected_distance;
-	float slice_hight;
+	double corrected_distance;
+	double slice_hight;
 
 	corrected_distance = fishbowl(info, distance, whichray); // il faut appliquer cette formule pour tout angle ou juste pour les angle droit??
-	// // printf("---------------------------------------------------------corrected_distance %f\n", corrected_distance);
+	// // // // printf("---------------------------------------------------------corrected_distance %f\n", corrected_distance);
 	slice_hight = projected_slice_hight(info, corrected_distance);
 	return (slice_hight);
 }
@@ -22,13 +22,13 @@ float walls(t_info *info, float distance, int whichray)
 **pour afficher un mur droit quand on est face a un mur
 */
 
-float fishbowl(t_info *info, float distorted_distance, int whichray)
+double fishbowl(t_info *info, double distorted_distance, int whichray)
 {
-	float angle_offset;
-	float corrected_distance;
+	double angle_offset;
+	double corrected_distance;
 
 	angle_offset = info->pov / info->screenWidth;
-	// // printf("angle_offset = %f\n", angle_offset);
+	// // // // printf("angle_offset = %f\n", angle_offset);
 	if (whichray < (int)(info->screenWidth / 2))
 		corrected_distance = distorted_distance * cos(ft_deg2rad((info->pov / 2) - (whichray * angle_offset)));
 	else
@@ -41,9 +41,9 @@ float fishbowl(t_info *info, float distorted_distance, int whichray)
 **adapter la taille du mur par rapport a la distance au plan et non la réelle distance
 */
 
-float projected_slice_hight(t_info *info, float distance)
+double projected_slice_hight(t_info *info, double distance)
 {
-	float projected_slice_hight;
+	double projected_slice_hight;
 
 	projected_slice_hight = (info->blocksize / distance) * info->ad;
 	return(projected_slice_hight);
